requirements.txt
certstream==1.12
PyYAML==6.0.2
wordsegment==1.3.1